requirements.txt
pymongo==4.6.2
instaloader==4.12.1
requests==2.31.0
python-dotenv==1.0.1